+++ src/apis/confirmOrder/confirmOrder.ts
@@ -1,7 +1,8 @@
+import { fetchData } from '../';
 import { generateBasicToken } from '../../utils';
+
 import { API_ROUTE, USER_ID, USER_PASSWORD } from '../constants/apiSetting';
 import { ERROR_MESSAGE } from '../constants/errorMessage';
-import fetchData from '../fetchData/fetchData';
 
 export async function submitOrder(cartItemIds: number[]): Promise<void> {
   const token = generateBasicToken(USER_ID, USER_PASSWORD);
