+++ src/components/confirmOrder/OrderContainer/OrderContainer.tsx
@@ -1,11 +1,10 @@
 import { useEffect } from 'react';
-
-import { CouponContainer, OrderItemContainer, RemoteShippingCheckBox } from '..';
-import { OrderAmount } from '../../common';
-
-import * as Styled from './OrderContainer.style';
-import { appliedCouponIdsState, discountAmountState } from '../../../recoil/atoms';
 import { useSetRecoilState } from 'recoil';
+import { appliedCouponIdsState, discountAmountState } from '../../../recoil/atoms';
+
+import { CouponContainer, OrderItemContainer, RemoteShippingCheckBox } from '../';
+import { OrderAmount } from '../../common';
+import * as Styled from './OrderContainer.style';
 
 export default function OrderContainer() {
   const setDiscountAmount = useSetRecoilState(discountAmountState);
